Stop using the server
