+++ client/public/electron.js
@@ -1,4 +1,8 @@
-const { app, BrowserWindow } = require('electron');
+const { app, BrowserWindow, globalShortcut, ipcMain } = require('electron');
+
+// const express = require('express');
+const fs = require('fs');
+// const bodyParser = require('body-parser');
 
 const path = require('path');
 const isDev = require('electron-is-dev');
@@ -15,7 +19,42 @@ app.on('activate', function() {
   if (mainWindow === null) createWindow();
 });
 
+ipcMain.on('readDirectory', () => {
+  // directoryContents('./');
+  console.log('--------------------')
+});
+
+function readFile(path) {
+  return fs.readFileSync(path, 'utf8');
+}
+
+function directoryContents(path) {
+  return fs.readdirSync(path);
+}
+
 function createWindow() {
+  // const app = express();
+  // // kill this port
+  // const port = process.env.PORT || 5000;
+
+  // app.use(bodyParser.json());
+
+  // app.post('/readFile', (request, response) => {
+  //   const { file } = request.body;
+
+  //   response.send({ fileContents: readFile(file) });
+  // });
+
+  // app.post('/readDirectory', (request, response) => {
+  //   const { path } = request.body;
+
+  //   response.send({ contents: directoryContents(path) });
+  // });
+
+  // app.listen(port, () => console.log(`Listening on port ${port}`));
+
+  //-------------------------------------------
+
   mainWindow = new BrowserWindow({
     width: 1400,
     height: 900
